Crée un composant générique pour afficher une liste de fiches action avec filtres
