+++ packages/ui/src/design-system/Modal/Modal.tsx
@@ -41,6 +41,13 @@ export type RenderProps = {
   ref: React.MutableRefObject<HTMLElement>;
 };
 
+export type ModalOpenState = {
+  /** état d'ouverture de la modale */
+  isOpen: boolean;
+  /* accompagne "isOpen" afin de pouvoir fermer la modale */
+  setIsOpen: (opened: boolean) => void;
+};
+
 /** Types des props du composant générique Modal */
 export type ModalProps = {
   /** Fonction de rendu du contenu de la modale */
@@ -58,12 +65,7 @@ export type ModalProps = {
   /** Titre et description centrés par défaut */
   textAlign?: 'left' | 'center' | 'right';
   /** Permet de contrôler l'ouverture de la modale */
-  openState?: {
-    /** état d'ouverture de la modale */
-    isOpen: boolean;
-    /* accompagne "isOpen" afin de pouvoir fermer la modale */
-    setIsOpen: (opened: boolean) => void;
-  };
+  openState?: ModalOpenState;
   /** fonction appelée lors de la fermeture de la modale */
   onClose?: () => void;
   /** max-width prédéfinies dans le DSFR, valeur par défaut "md" */
